Revert "API: Add Compliance API slice"
This reverts commit 36f9e7056543f2e780b46b4d8e100ee33324bb7e.

The Compliance v1 API is not intended to be consumed by teams other than
Compliance and is not stable. For now, we will implement an OpenSCAP MVP
without calling the Compliance API. The Compliance team plans to release
a stable, consumable v2 API in the near future and we will consume that
when it is ready.